pyproject.toml
[project]
name = "limberer"
version = "0.6"
authors = [
  { name = "Jeff Dileo", email = "jtdileo@gmail.com" },
]
maintainers = [
  { name = "Jeff Dileo", email = "jtdileo@gmail.com" },
]
license = { file = "LICENSE" }
description = "A flexible document generator based on weasyprint, mustache templates, and pandoc."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["pandoc", "panflute", "markdown", "html", "pdf", "report", "document", "generator"]
classifiers = [
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3.9",
]
dependencies = [
  "panflute",
  "chevron",
  "weasyprint",
  "rtoml",
  "pillow",
  "beautifulsoup4",
  "Pygments",
]

[project.urls]
repository = "https://github.com/ChaosData/limberer"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
limberer = "limberer:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"limberer.static" = ["*", "assets/*", "custom/*", "sections/*", "templates/*", "images/*"]
